requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2